drm/bridge: ti-sn65dsi86: support booloader enabled display
Request the enable gpio ASIS to avoid disabling bridge during probe, if
already enabled.  And if already enabled, defer enabling runpm until
attach to avoid cutting off the power to the bridge.

Once we get to attach, we know panel and drm driver are probed
successfully, so at this point it is safe to enable runpm and reset the
bridge.  If we do it earlier, we kill efifb (in the case that panel or
drm driver do not probe successfully), giving the user no way to see
what is going on.

Signed-off-by: Rob Clark <robdclark@chromium.org>